sample_project/requirements.txt
-e svn+http://django-db-log.googlecode.com/svn/trunk/#egg=djangodblog
django-extensions
